Temporary fix for Keldoc timeouts.
Some API calls on three or four vaccination centers are very slow (sometimes more than 3 minutes). Instead of cutting off the possibility to retrieve the list of Keldoc centers, it would be better to put a TImeout. The execution time of Keldoc is more reasonable (30-40 seconds on my side).

However, Keldoc uses an unoptimized way to handle vaccination slots. Indeed, on some centers, it is sometimes necessary to make a dozen requests on the agendas because it is necessary to retrieve information from several sub-agendas (not listed on their site) which are attached to different medical teams, which can slow down the process.

Until a solution is found to retrieve the next available date information with fewer queries (if that is possible?), using a timeout will sacrifice two to three centers that take a long time to give results.

I am still currently working on this issue.